US53 - JR/SM - Adds the activity and the presenter for PTV Report

+++ app/src/main/java/org/openlmis/core/model/ServiceDispensation.java
@@ -33,7 +33,7 @@ import lombok.Setter;
 public class ServiceDispensation extends BaseModel implements Serializable {
 
     @DatabaseField
-    int quantity;
+    long quantity;
 
     @DatabaseField
     String signature;
